+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/games/CallbackGame.kt
@@ -3,8 +3,4 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.games
 import kotlinx.serialization.Serializable
 
 @Serializable
-class CallbackGame {
-    init {
-        TODO()
-    }
-}
+object CallbackGame